Update plugin: gravityforms.
Old Version: 2.10.0
New Version: 2.10.3

See #3718.

+++ wp-content/plugins/gravityforms/forms_model.php
@@ -2507,6 +2507,27 @@ class GFFormsModel {
 		 */
 		$file_path = apply_filters( 'gform_file_path_pre_delete_file', $file_path, $url );
 
+		// If the file URL is outside the uploads folder, something nefarious is up, so bail.
+		if ( ! GFCommon::is_file_in_uploads( $url ) ) {
+			GFCommon::log_debug( __METHOD__ . sprintf( '(): Not deleting file from URL: %s', $file_path ) );
+			return;
+		}
+
+		// Verify the resolved file path is within the uploads root (defense-in-depth against filter manipulation).
+		$resolved_path = GFCommon::get_absolute_path( $file_path );
+		$upload_root   = trailingslashit( GFCommon::get_absolute_path( self::get_upload_root() ) );
+		if ( ! str_starts_with( $resolved_path, $upload_root ) ) {
+			GFCommon::log_debug( __METHOD__ . sprintf( '(): Not deleting file; resolved path is outside the uploads root: %s', $file_path ) );
+			return;
+		}
+
+		// if file path contains traversal characters or null bytes, something nefarious is up, so bail.
+		$path_validation = GF_Download::validate_file_path( $file_path );
+		if ( is_wp_error( $path_validation ) ) {
+			GFCommon::log_debug( __METHOD__ . sprintf( '(): Not deleting file (%s): %s', $path_validation->get_error_code(), $file_path ) );
+			return;
+		}
+
 		if ( file_exists( $file_path ) ) {
 			$result = unlink( $file_path );
 
@@ -8565,7 +8586,8 @@ class GFFormsModel {
 	 * $_POST['gform_uploaded_files'] and caches them in GFFormsModel::$uploaded_files.
 	 *
 	 * @since 2.4.3.5
-	 * @since 2.9.18 Deprecated the string-based (file/basename) input value. Added support for dynamically populated file URLs using the `url` key.
+	 * @since 2.9.18 Added support for dynamically populated file URLs using the `url` key.
+	 * @since 2.10.3 Deprecated the string-based (file/basename) input value.
 	 *
 	 * @param int $form_id The ID of the form the submission is being processed for.
 	 *
@@ -8591,6 +8613,12 @@ class GFFormsModel {
 							continue;
 						}
 
+						if ( isset( $file['url'] ) ) {
+							GFCommon::log_debug( __METHOD__ . sprintf( '(): Removing Url %s. File uploads must be submitted as binary uploads.', $input_name ) );
+							unset( $input_files[ $key ] );
+							continue;
+						}
+
 						// All files regardless of upload or population method should have this.
 						if ( isset( $file['uploaded_filename'] ) ) {
 							$file['uploaded_filename'] = sanitize_file_name( wp_basename( $file['uploaded_filename'] ) );
@@ -8601,11 +8629,6 @@ class GFFormsModel {
 							$file['temp_filename'] = sanitize_file_name( wp_basename( $file['temp_filename'] ) );
 						}
 
-						// Used when the field is dynamically populated on initial form display.
-						if ( isset( $file['url'] ) ) {
-							$file['url'] = esc_url_raw( $file['url'] );
-						}
-
 						// Sanitize or generate the UUID to be used by the file preview and error messages markup.
 						if ( isset( $file['id'] ) ) {
 							$file['id'] = sanitize_key( $file['id'] );
@@ -8613,8 +8636,19 @@ class GFFormsModel {
 							$file['id'] = GFFormsModel::get_uuid();
 						}
 					}
+
+					$input_files = array_values( $input_files );
+					if ( empty( $input_files ) ) {
+						unset( $files[ $input_name ] );
+					}
 				}
 			} else {
+				if ( GFCommon::is_valid_url( $input_files ) ) {
+					GFCommon::log_debug( __METHOD__ . sprintf( '(): Removing URL %s. File uploads must be submitted as binary uploads.', $input_name ) );
+					unset( $files[ $input_name ] );
+					continue;
+				}
+
 				// Deprecated, retaining for backwards compatibility with third-party integrations.
 				$input_files = wp_basename( $input_files );
 			}
